requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyserial==3.5
werkzeug==2.3.7